Removes bad symbolism (#2512)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes harmful symbolism from the game
from

![{7272D7F4-F9CE-4C90-A4E8-3E1B6188D70E}](https://github.com/user-attachments/assets/25279944-3922-4188-a4bd-7f593770949f)


to

![{B168D387-EB1A-4364-9CDE-1BA06605BC17}](https://github.com/user-attachments/assets/6c76ac28-8022-46e5-aac0-e454b5e46c42)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bad symbolism bad
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![{56C36B56-DC16-435F-BB03-620A956B20F3}](https://github.com/user-attachments/assets/e7a0781d-2d16-4d6f-a234-fff77939bb20)<!--
Show proof of testing, screenshots or recordings for features, proof of
compile for backend changes -->