Refugee merchant update (#2429)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

adds some features that were lacking from the original refugee pr, also
give spicer a statblock.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

there is zero reason to play the refugee merchant currently, this adds a
reason and some variety trade wise.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/cf7634d4-8fee-4a93-9c7d-778b602633a2)

![image](https://github.com/user-attachments/assets/0beb241c-f199-48a4-aa93-024714b58d05)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->