Wanted bandits (#2492)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
**This PR need to be TM'ed for very clear balance reasons.**

This adds the "WANTED" trait, a optional play style for bandits that
would rather start out as wanted, but with more skills.
This should not be a straight up "buff" for bandits, but the bandits
choosing to have a "curse" to get better stats.

This helps with a few things: 
1. Bandits right now are usually seen as very boring as passive, this
was introduced when the bandit tag was removed, for better or worse.
2. A lot of keep players have a bad problem with meta, it's a huge
problem that came up after the bandit tag was removed.
3. We are adding more play styles for the bandits
4. We are introducing the gate having to check people faces at the gate,
something, that was kind of removed and left the vanguard with only
"asking" for coin, and not doing anything else at the gate.

In terms of stats:
If you choose not to be wanted: You get a poor pouch with coins.
Something, that was lacking beforehand so they could not enter the town.

A change for the hedgeknight will be made, as he will be wanted by
default now and instead of 30 PQ then it will be 15 PQ (still a lot of
PQ, but not insanely so)
Rogue Mage will be getting armor by default (something he have not had
for some reason) and a leather belt
```
armor = /obj/item/clothing/suit/roguetown/shirt/robe/newmage/adept
belt = /obj/item/storage/belt/rogue/leather
```
If wanted: 
brigand: 
```
axes, 1
maces, 1
whipsflails, 1
unarmed, 1
```
foresworn
```
polearms, 1
wrestling, 1
unarmed, 1
swords, 1
crossbows, 1
```
knave
```
endurance 2
constitution 1
sneaking, 1
athletics, 1
```
sawbone
```
knives, 1
swords, 3
wrestling, 1
athletics, 1
treatment, 1
```
rogue mage
```
backpack_contents = list(/obj/item/book/granter/spellbook/expert = 1)
intelligence 2
arcane, 1
```
hedge knight
```
intelligence 1
speed 1
```

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More options for play styles
Introducing more shenanigans that should make for less boring rounds
Adding more for the roles to do, while it being "important" to check if
people are bandits again
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
![Skærmbillede 2025-03-01
150921](https://github.com/user-attachments/assets/b3d63f5d-f29a-4ca0-bc8c-da9c698895f0)
![Skærmbillede 2025-03-01
134912](https://github.com/user-attachments/assets/893ee33d-baab-4720-ad5a-828a49d88f45)
![Skærmbillede 2025-03-01
134955](https://github.com/user-attachments/assets/997812c4-a390-430c-a280-5b2fc5fbe9f2)

https://youtu.be/QSSCM68oApI
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

---------

Co-authored-by: EvenInDeathIStillServe <snoopdogog@gmail.com>