Refugee buffs (#2348)
Refugee buffs

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR buffs refugees, making them a little closer to on par with
towners, but still being weaker then them. Currently they are just
abysmally bad

For the main part refugees get a vest and a coif for basic protection
against NPCs, eventually i want to learn spriting and add in makeshift
armor for refugees

May also add in a couple of new refugees throughout the PR

## Why It's Good For The Game

Refugees are practically useless, their skills are abysmal for people
that apparently survived trekking long distances through a zombie
apocalypse.


## Proof of Testing (Required)

![image_2025-02-01_045309293](https://github.com/user-attachments/assets/abf15748-8159-4adb-90f8-d0960a34ebd5)