The TerrorBog- Unfamiliar Terrain (#2476)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR creates a debuff called 'unfamiliar terrain' applied to every
guard role with the trait 'bog vulnerable'. (newly added). (Now
renamed!)
This includes knights, MAA, city watch, gate keeper, and Guard captain.
Noticably, the Vanguard is explicitly exempt from this debuff.
This debuff applies when a guardmember with the guardsmen trait moves
into a 'bog' area; Currently, past the old bog gate.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
For as long as I have started playing on ratwood, we've always had an
issue where someone reports a bandit, or one of our antags in the bog.
The keep immediately rallies pretty much every combat role, even those
meant to guard the keep proper, and then sends them out into the bog. A
lot of times, the guards will do it unprompted from orders of the duke.
This PR should /hopefully/ reduce it's frequency.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/86a6ad8a-7272-496c-9374-3ffa1e9d5e3d)

![image](https://github.com/user-attachments/assets/a0846abf-2b95-4851-bb61-81284df8d97b)

![image](https://github.com/user-attachments/assets/21b72703-6f39-422a-9c93-c5361ccd9618)

![image](https://github.com/user-attachments/assets/78533aa9-c307-4b39-bb23-4d712b0cfd31)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

---------

Co-authored-by: EvenInDeathIStillServe <snoopdogog@gmail.com>