Fixes specials flavour text, removes bad specials (#2507)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes specials text to be a lot more sensible.

Removes some specials outright bad for the game.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
A lot of the old specials were either narcissistic or straight up
referring to things that don't exist??? or just plain bad for roleplay
and immersion by setting a bad precident.

Removed specials:

Nothing

Imagine rolling and getting told "fuck you, no playing."

Master craftsman & Sword master & Pineapple
The powergamers pick, take master craftsman on a templar, steward, duke
etc and fuck over all the towners or take sword master on mage or towner
to do the same but inverse.

Armour training
Just play any role that has armour training.


Mule & Merchant
fuck the merchant and fuck the nightmaster if it was a special.

Thinker
Similar issue to sword master and jack of all trades except it's more
mages in a game that does not need more mages.

Warhorse

No, veteran does not need a buff, stop asking.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

https://cdn.discordapp.com/attachments/1240810578029576292/1347966659155202138/image.png?ex=67cdbf47&is=67cc6dc7&hm=96fc4765eeb2137bf3f2cd84ef50afaacaf34aad6bda12ecbf5343ca4fae56b3&
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->