Stalker Bow Choice (#2569)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Let's you choose between a recurve bow and a crossbow for the rogue
stalker
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The stalker class was made to be the new bow focused combat class for
refugees after they were all removed with the refugee pr. We already
have 2 other classes with crossbows, the deserter crossbowman and in the
same rogue file as the stalker, the smuggler.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/cc4ab04b-57e7-41e0-bc41-c4a4e71e9d40)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->