Fixes the carriage collision bounds. (#2572)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
You can run under the carriage. This expands the bounds of the carriage
so you can't do that anymore.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People will abuse this and hide under it making them impossible to hit.
It also seems unintentional.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/8396692a-df5b-40b2-935f-e52b7e5e4401


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->