Refugee wave removal (#2570)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A unlimited amount of refugee can be spawned which means that this wave
was never used outside of events. The code was changed so that a refugee
wave could be spawned in by a admin, while removing a wave that nobody
really used.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Refugee waves are not used by anyone, and people have their time wasted
while refugee wave that nobody plays goes by them.
If someone really wanted a refugee wave, then it's usually because of a
event.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![billede](https://github.com/user-attachments/assets/9c14f84e-dad7-4204-a562-a3039b2feddc)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->