Kneestinger immunity to Druid (#2530)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Simple one line change that adds the kneestinger immunity traits to
druids.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

This was originally removed from druid in a very controversial PR with
nineteen downvotes, and only three upvotes.
([#960](https://github.com/Rotwood-Vale/Ratwood-Keep/pull/960)). Druids
no longer can summon kneestingers, and it's not very abuseable except
against simplemobs, which can be killed with kneestingers even without
this trait. The kneestinger immunity trait is also going unused at the
moment which is sad because it's a flavorful inclusion for a druid.

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

![image](https://github.com/user-attachments/assets/5c3ceafc-80c8-45a0-a255-bcc532406e1b)