Decapitation fixes (#2574)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes mind transfer to brain
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This code was commented out by azarak 9 months ago, and re-enabled by
gyran during his cleanup PRs. This is what causes the mind to transfer
to the brain rather then remain in the body. Removing should fix
majority of decapitation issues.

This fixes the root problem that  #2565 is trying to address.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/e63e3f9c-e8a4-4cab-9a5c-c5a5dfc3903f)

![image](https://github.com/user-attachments/assets/01cf8288-ac77-4508-be9b-b9a6ce54bf6d)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->