Adds vanguard cloak sprites for small races (#2566)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
They had missing sprites, this adds them and also fixes the naming
convention of the one currently present. As for the 'xl_' versions: I
can't manipulate the cloaks in any way to make it use them so I ignored
those for now. If anyone knows how and when they're used I can test
those too but it doesn't seem like they're used currently.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Small races can wear these but they don't show up. Saw it disappoint
like 3 people in a single round.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Little hard to see (bad example sorry) but it's there.

![cloak](https://github.com/user-attachments/assets/8aa4758f-8987-4874-b426-f62bba2a47a9)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->