Fishing cages and shellfish! [Bounty] (#2522)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR adds two main new things, the fishing cage, which after
deploying it on a body of water and putting some bait on it, will catch
something. And the new stuff it can catch, shellfish! I'll simply make a
list for easy reading down below:

Note: I'll be leaving on [] the names of the persons that have made the
sprites for the items. These are their names on the discord, you can
probably find them on the thread of the bounty here too:

- Fishing cages! [kino1208]
-Craftable with a log and two sticks
-Has a different loot table, so they have shellfish instead of fish, and
a lower chance to catch junk.
-Time to catch depends on skill, no skill and novices taking 20 minutes,
apprentices at 15, journeymen at 10, experts at 5, masters at 3, and
legends on only 90 seconds to catch.
-Raises fishing skill, so you can indeed train it up by setting a bunch
of different fishing cages.
-Carriable on your back slot

![image](https://github.com/user-attachments/assets/ac13284a-b72b-442f-b816-f44a832370cd)


- New shellfish!
-Shrimps are the equivalent of eels and they have new and more detailed
sprites [Fridge]
-Oysters are the equivalent of carps, you can pry them open with a knife
and, if you're lucky, get a pearl inside them. The rarer the oyster, the
higher the chance of a pearl being inside (right click to take it). If
you're extra lucky, you may even get a black pearl. After opening them,
you can use a knife again to take the meat out and get empty oyster
shells. Or just eat the raw meat if you're like that. [ExcusA]
-Crabs are the equivalent of anglerfish [Fridge]
-Lobsters are the equivalent of clownfish [Fridge]

![image](https://github.com/user-attachments/assets/692c99a0-4e87-4709-9d5d-c76b89f78dc3)

- New valuables!
-Adds four new valuable items:
-Pearl amulet, needs 3 pearls and 1 fiber to craft, can be used for
praying and casting [ExcusA]
-Black pearl amulet, needs 3 black pearls and 1 fiber, extra rare
version of the pearl amulet [ExcusA]
-Shell necklace, needs 5 oyster shells and 1 fiber, can be used for
praying and casting too [Fridge]
-Shell bracelet, same as above, but needs 3 oyster shells, and can only
be worn on the wrists. [Fridge]


![image](https://github.com/user-attachments/assets/32f794ef-0eb3-4caf-b107-706c3c9ab31c)
![image](https://github.com/user-attachments/assets/0b9b4063-5d59-428a-a71a-b333b9ed0cc2)

- New foods!
-Adds four different new meals that can be made, one for each type of
shellfish:
-Crab cake, made by simply taking a raw crab and covering it with small
dough, and then throwing it on the oven or frying it for a bit. [Fridge]
-Lobster meal, made with a slice of butter and a cooked lobster, pretty
much the fish equivalent of a peppersteak. [Fridge]
-Shrimp gumbo, made with two cooked shrimps, some cabbage and an onions
slice with a bowl, very filling since it's harder to make (Crafting
menu) [patches1775 (Dame Anya Dawnheart)]
-Oyster chowder, made with one opened oyster, some powder (flour), and a
sliced onion with a bowl, also very filling and tasty (Crafting menu)
[patches1775 (Dame Anya Dawnheart)]

![image](https://github.com/user-attachments/assets/69d851b5-1fd4-4c78-9c22-333fbbde175e)
![image](https://github.com/user-attachments/assets/86da6a9e-1931-433a-9478-13e90dd195bc)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More stuff for the fishers to do, more fishes in general, more items for
abyssor worshipping and more dishes to make from sea creatures. Pretty
much more things to do as a fisherman in general.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
https://discord.com/channels/1240735983276654733/1322035985592225896
+ all of the pictures above
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->