Death Consequences 2.0 (#2537)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Death consequences but a bit more forgiving giving you an additional
chance, even if severely crippled.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Death consequences got removed with a pr that removed the standard ways
of reviving people and were not readded when the pr was mostly reverted,
this brings it back. Also it's weird that you can be revived 5 times and
still be fully combat capable.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/d388ecb2-8f3c-46c3-8113-a11b6e1c6033)

![image](https://github.com/user-attachments/assets/59ba673e-4174-4d48-ac0b-eee0998ed8d0)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->