Bottle Bomb: lowers chance of fail, fixes qdel issue. (#2446)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Bottle bomb for exploding is dreadful. What this does

- Lowers the default probability of failure to 1 (1%?)
- Moves the qdel around so it only deletes when it explodes or smashes
apart.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
The bottle bomb's rate of failure, along with its ability to delete
itself without any warning frustrates many. It's a problem. This makes
them a lot more likely to succeed.

Should a bottle bomb fail to explode it will lay on the floor defused.
This will allow you to try again or anyone else to pick it up, however
the fuse is now a lot shorter so be careful! There is also a chance
(very rare) the bottle bomb will snuff out in your hand. In this case
the bottle bomb no longer has a fuse and would ignite on being exposed
to a flame. This would be a pretty rare circumstance but I also think it
would be funny to see that so I left it in.

## Proof of Testing (Required)


https://github.com/user-attachments/assets/f3e6b957-ca66-44dd-8af7-c7d0836bbd54


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->