Removes goblin camp immunity to being found (#2371)
## About The Pull Request

This pull request is a pretty simple one- it removes the immunity of the
goblin camp from being found by people who witness a goblin entering the
camp, or are dragged in without a hood on.

## Why It's Good For The Game

The goblins possess two entrances into the town, as well as one right in
the lowtown. They are given a very good position to be able to kidnap
people- and have three entry points that they can use to avoid pursuit.
Goblins regularly kidnap people- far moreso than bandits are able to,
due to their unique advantage with the positioning of their entrances. I
have seen goblins abuse this- for example, a Duchess being kidnapped
with people actively present and right there to see the goblin go
through- only for the goblin to manage to get the travel tile to pop,
leaving the entire retinue stood there with their thumbs up their asses,
with no way to retaliate whatsoever, despite having seen the goblin go
through.

It's safe to say that's not fair. Goblins already have the advantage of
having three entrances and exits- they can always attempt to lose their
tail if they're being pursued, and this will encourage them to act more
smartly rather than relying on their ability to completely lock out
anyone from being able to pursue them.

## Proof of Testing (Required)

https://gyazo.com/a757b16ccf3dee42e1748774a20e9961

Compiled without any errors.