Goblin Prisoner and Unbanditifying Prisoner Roles (#2559)
## About The Pull Request

Allows goblins to be bog and rockhill prisoners.
Removes the bandit camp trait from bog and rockhill prisoners.
Ensures goblins spawn with graggar as patron.

## Why It's Good For The Game

No real reason to not let goblins roll prisoner.
Giving prisoners the bandit camp trait is just an opening for gamey
exploitation which has occurred already in the past. It also forces the
characters to background-wise be bandits after all.


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/bfb15332-6609-490b-981b-fbfffd4429fc)

![image](https://github.com/user-attachments/assets/6d28a1bb-d258-492a-8e78-41a996ce4e08)