Tavern map changes (#2452)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR changes the tavern to be more of a social place, rather than..
whatever it was supposed to be before. It adds more tables small where
people can sit around and talk with eachother, adds some cards and dice
for people to play, adds a couple of tables outside so people can sit
around there while watching whats happens around on the street or
watching the local execution, adds a couple of windows so people can
take a peek inside or look at what's happening outside without having to
leave or enter, adds shutters to the bar for when the barkeep isn't
around, and changes the kitchen layout, so noticing when a cook is
around is easier and the cook can more easily put multpile dishes
without struggling to fit it all in one table.

Photos are on the proof of testing.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Hopefully it'll make the tavern more of a social place where people can
hang around and chat, like it was before, giving more chances to
roleplay and do some activities with the dice and cards.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/172d3247-8569-442f-a9ae-aeda04070988)

![image](https://github.com/user-attachments/assets/258d3425-21b2-47de-a72f-f115c63c8573)

![image](https://github.com/user-attachments/assets/6bba86fc-e075-4a62-a407-04eecc4d2110)

![image](https://github.com/user-attachments/assets/7fd5c52d-f22c-4718-bbf6-97914b500cc3)

![image](https://github.com/user-attachments/assets/2906ce31-91f2-4458-9587-0acac3b3d751)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->