New set of dices to GAMBLE (#2502)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds new bag of dice to play a particular dice game (yes, from KCD)
Adds a couple of these bags to the map and to the luxury list of
shopkeep
Gave the bag to the drunkard refugee type
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Lets play fucking dice

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/351542dc-6e0d-48db-a312-7e5fc84d6d11)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->