Reworks platter code (#2580)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
(Updated Re-submission of
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2549)
Platters currently only work for certain foods, they have specific
sprites, and then return specific platters on consumption. This is
disastrous as they only return wooden platters when other types exists
meaning your old platter will forever become a wooden one once used.

This PR changes the code to allow any foods to be put on a platter and
does the following:
- Creates a datum of sprites to use when certain types are used on
platters. This is an associated list which uses the path type as a key
and the platter sprite's name as the value.
- Allows any food/snacks subtype to be allowed into a platter's
contents.
- uses this contents item to create a vis_contents for the platter.
- renames the platter to be 'platter of x'.
- gives the platter the item's description.
- Adds the bonus nutriments platter food gets to any placed on platters
(probably needs to be changed to a buff or something, currently it's the
only incentive to use platters outside of RP).
- Right clicking will allow you to remove anything on the platter (so
long as your hand is free). Also removes bonus nutriment
- comments out all the previous attackby code for platters
- adds platter sprites for previously existing sprites which had them
(removes platter from sprite)
- Adds platter to the list of things that prevent food rot (tables
prevent it too?? Seems like only chests should but it can be changed
later if needed.)

I tested this quite a bit and haven't been able to find anything wrong
with it but I'm willing to fix any issues it produces in the future if
someone lets me know about them.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Platters simply don't work outside of weird manually coded exceptions
and even then they break the subtypes of platters. This will allow
people to put any single food on platters, get a small benefit from it,
and keep the platter they started with.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/3b3b1f98-a7de-4c4e-891e-5f75ceeab0c7


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->