Fix for over/underflowing stat boosts causing round-permanent stat changes (#2587)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR fixes an issue where applying stat boosts or reductions which
take a stat above or below the maximum or minimum values (20 and 1,
respectively) could sometimes cause the stat to permanently change after
the stat change is removed.

As part of the fix, I've also taken the time to refactor the code - the
same code was duplicated for each stat, while the new code is wrapped in
a single proc. I've also removed an unused "statbuf" variable.

One detail is that I've added a variable, "tempskill". This variable is
of type /obj/statdata, which is defined at the end of the file, to allow
code reuse rather than having to duplicate the stat change code for each
stat. In principle, this could (probably? I don't know the details of
BYOND's threading model) be replaced with a single global piece of data,
since it's only used to allow the stat change code to return two
variables at once. Please let me know if there's a better way of
achieving this effect and I'll be happy to change that detail of the PR.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The previous stat boost logic could cause stat reductions for the rest
of the round if a player wasn't careful with which buffs they were
applying and when, which could lead to frustration - a friend of mine
mentioned his intelligence becoming too low for certain magos activities
as a result of the issue, for instance. This fix prevents that
frustrating situation.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Image prior to any buffs, showing intelligence at 19:

![image](https://github.com/user-attachments/assets/106153c4-0093-45dd-b946-ec7da0ce0db4)


Image captured following use of spice, showing intelligence at 20. If
not for the cap, it would instead to 24:

![image](https://github.com/user-attachments/assets/c80c421f-0c14-4354-9426-122518486883)

Image captured after additionally taking swampweed, boosting
intelligence by another 2, showing intelligence still at 20, plus an
image showing the background BUFINT variable, which stores the amount we
currently exceed the limit by (6).

![image](https://github.com/user-attachments/assets/bf82decd-d1b1-411a-b9bb-e602638177cd)

![image](https://github.com/user-attachments/assets/8200fba8-3365-42dd-9f7c-0eb6fa4f88b1)

Image after spice has worn off, showing BUFINT of 1, and intelligence
still at 20:

![image](https://github.com/user-attachments/assets/d9da876d-2a38-47bb-b6a5-250ea07e56f2)

Finally, after both buffs have worn off, intelligence is back to its
original value of 19:

![image](https://github.com/user-attachments/assets/eefdee91-cf58-4c3d-89ab-3981382a9fdc)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->